// src/app/core/models/curso.interface.ts
export interface Curso{
    nombreProfesor: string;
      genero: string;
      id: string;
      telefono: number;
      nombreCurso: string;
      idCurso: string;
      fecha: Date;
      email: string;
      tiempo: string;
      precioCurso: number; 
      tipoCurso: string;
      salon: string;
      Descri: string;
  }
  